requirements.txt
alabaster==0.7.12
awscli==1.18.155
Babel==2.8.0
botocore==1.18.14
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
country-converter==0.7.0
coverage==5.3
cycler==0.10.0
decorator==4.4.2
docutils==0.16
flake8==3.8.4
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.2.0
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.2
matplotlib-venn==0.11.5
mccabe==0.6.1
networkx==2.4
numba==0.51.2
numpy==1.19.2
packaging==20.4
pandas==1.1.3
Pillow==7.2.0
pyarrow==0.17.0
pyasn1==0.4.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
-e git+git@github.com:jaimeoliver1/social_capital_in_trade_networks.git@ce4b174e9c7e4094417cb2f2f20d0bcbc592bae7#egg=src
threadpoolctl==2.1.0
umap-learn==0.4.6
urllib3==1.25.10
zipp==3.3.0
